+++ src/hasher/hamdist.c
@@ -30,7 +30,9 @@ errors that transformed one string into the other.
 
 #endif
 
-static int hamdist(int x, int y)
+#include "../radare.h"
+
+static int hamdist(int x, u64 y)
 {
 	int dist = 0, val = x^y;
 
@@ -42,7 +44,7 @@ static int hamdist(int x, int y)
 	return dist;
 }
 
-unsigned char hash_hamdist(char *buf, int len)
+unsigned char hash_hamdist(char *buf, u64 len)
 {
 	int x = 0;
 	int y = 0;
